refactor: 카카오톡 공유 로직 수정 및 템플릿 아이디 상수화
