+++ frontend/src/constants/kakao.ts
@@ -1,0 +1,2 @@
+export const MEETING_INVITE_TEMPLATE_ID = 110853;
+export const MEETING_RESULT_TEMPLATE_ID = 111325;
